Cambios en Mi Perfil
Linea de tiempo de usuarios lista

+++ Controllers/UsuarioController.php
@@ -1,7 +1,9 @@
 <?php
 include_once '../Models/Usuario.php';
+include_once '../Models/Historial.php';
 include_once '../Util/Config/config.php';
 $usuario = new Usuario();
+$historial = new Historial();
 session_start();
 
 if ($_POST['funcion'] == 'login') {
@@ -82,25 +84,54 @@ if ($_POST['funcion'] == 'editar_datos') {
     $email = $_POST['email_mod'];
     $telefono = $_POST['telefono_mod'];
     $avatar = $_FILES['avatar_mod']['name'];
-    if($avatar!=''){
-        $nombre = uniqid().'-'.$avatar;
-        $ruta = '../Util/Img/Users/'.$nombre;
-        move_uploaded_file($_FILES['avatar_mod']['tmp_name'],$ruta);
-        $usuario->obtener_datos($id_usuario);
-        foreach ($usuario->objetos as $objeto){
-            $avatar_actual = $objeto->avatar;
-            if($avatar_actual!='user_default.png'){
-               unlink('../Util/Img/Users/'.$avatar_actual);
-            } 
+    $usuario->obtener_datos($id_usuario);
+    $datos_cambiados = 'Ha echo los siguientes cambios: ';
+    if($nombres!=$usuario->objetos[0]->nombres||$apellidos!=$usuario->objetos[0]->apellidos||$rut!=$usuario->objetos[0]->rut||$email!=$usuario->objetos[0]->email||$telefono!=$usuario->objetos[0]->telefono||$avatar!=''){
+        
+        if($nombres!=$usuario->objetos[0]->nombres){
+            $datos_cambiados .= 'Su nombre cambió de '.$usuario->objetos[0]->nombres.' a '.$nombres.', ';
         }
-        $_SESSION['avatar'] = $nombre;
-    }else{
-        $nombre = '';
-
-    }
+        if($apellidos!=$usuario->objetos[0]->apellidos){
+            $datos_cambiados .= 'Su apellido cambió de '.$usuario->objetos[0]->apellidos.' a '.$apellidos.', ';
+        }
+        if($rut!=$usuario->objetos[0]->rut){
+            $datos_cambiados .= 'Su rut cambió de '.$usuario->objetos[0]->rut.' a '.$rut.', ';
+        }
+        if($email!=$usuario->objetos[0]->email){
+            $datos_cambiados .= 'Su email cambió de '.$usuario->objetos[0]->email.' a '.$email.', ';
+        }
+        if($telefono!=$usuario->objetos[0]->telefono){
+            $datos_cambiados .= 'Su telefono cambió de '.$usuario->objetos[0]->telefono.' a '.$telefono.', ';
+        }
+        if($avatar!=''){
+            $datos_cambiados .= 'Cambió su avatar, ';
+            $nombre = uniqid().'-'.$avatar;
+            $ruta = '../Util/Img/Users/'.$nombre;
+            move_uploaded_file($_FILES['avatar_mod']['tmp_name'],$ruta);
+            $usuario->obtener_datos($id_usuario);
+            foreach ($usuario->objetos as $objeto){
+                $avatar_actual = $objeto->avatar;
+                if($avatar_actual!='user_default.png'){
+                   unlink('../Util/Img/Users/'.$avatar_actual);
+                } 
+            }
+            $_SESSION['avatar'] = $nombre;
+        }else{
+            $nombre = '';
     
-    $usuario->editar_datos($id_usuario, $nombres, $apellidos, $rut, $email, $telefono, $nombre);
-    echo 'success';
+        }
+        
+        
+        
+        $usuario->editar_datos($id_usuario, $nombres, $apellidos, $rut, $email, $telefono, $nombre);
+        $descripcion = 'Ha editado sus datos personales, '.$datos_cambiados;
+        $historial->crear_historial($descripcion,1,1,$id_usuario);
+        echo 'success';
+
+    }else{
+        echo 'danger';
+    }
+   
 }
 
 
@@ -115,6 +146,8 @@ if ($_POST['funcion'] == 'cambiar_contra') {
         if($pass_bd==$pass_old){
             $pass_new_encriptada = openssl_encrypt($pass_new,CODE,KEY);
             $usuario->cambiar_contra($id_usuario, $pass_new_encriptada);
+            $descripcion = 'Ha cambiado su contraseña';
+            $historial->crear_historial($descripcion,1,1,$id_usuario);
             echo 'success';
         }else{
             echo 'error';
